+++ src/components/SiteTitle.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+import Logo from "./Logo";
+
+export default function SiteTitle() {
+  return (
+    <header>
+      <Logo />
+    </header>
+  );
+}